Consolidate store down to single file
