--- src/app/shared/stores/domain/_store/index.ts
@@ -1,3 +1,0 @@
-export * from './temp.query';
-export * from './temp.store';
-export * from './temp.service';
