--- src/app/shared/stores/domain/static/index.ts
@@ -1,3 +1,0 @@
-export * from './static.query';
-export * from './static.store';
-export * from './static.service';
