--- src/app/shared/stores/domain/static/static.query.ts
@@ -1,10 +1,0 @@
-import { Injectable } from '@angular/core';
-import { Query } from '@datorama/akita';
-import { StaticStore } from './static.store';
-
-@Injectable({ providedIn: 'root' })
-export class StaticQuery extends Query<StaticState> {
-  constructor(protected store: StaticStore) {
-    super(store);
-  }
-}
