+++ src/app/shared/stores/domain/static/static.service.ts
@@ -1,11 +1,30 @@
 import { Injectable } from '@angular/core';
-import { StaticStore } from './static.store';
 import { HttpClient } from '@angular/common/http';
-import { StaticQuery } from './static.query';
+import { Query, Store, StoreConfig } from '@datorama/akita';
+
+/** Store */
+@Injectable({ providedIn: 'root' })
+@StoreConfig({ name: 'static' })
+export class StaticStore extends Store<StaticState> {
+  constructor() {
+    super({ todos: [] });
+  }
+}
+
+/** Query */
+// tslint:disable-next-line:max-classes-per-file
+@Injectable({ providedIn: 'root' })
+export class StaticQuery extends Query<StaticState> {
+  constructor(protected store: StaticStore) {
+    super(store);
+  }
+}
+
 /**
  * The simple domain store is for web api requests that are GET only and do not need create/update/delete
  * This is to eliminate the need to have a new store for every web api call
  */
+// tslint:disable-next-line:max-classes-per-file
 @Injectable({ providedIn: 'root' })
 export class StaticService {
   public todos$ = this.query.select(state => state.todos);
