--- src/app/shared/stores/domain/users/index.ts
@@ -1,3 +1,0 @@
-export * from './users.query';
-export * from './users.store';
-export * from './users.service';
