+++ src/app/shared/stores/domain/users/users.service.ts
@@ -2,14 +2,35 @@ import { Injectable } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
 import { tap, catchError, take } from 'rxjs/operators';
 import { throwError } from 'rxjs';
-import { applyTransaction } from '@datorama/akita';
+import { applyTransaction, EntityState, EntityStore, QueryEntity, StoreConfig } from '@datorama/akita';
 
-import { UsersStore, uniqueId } from './users.store';
-import { UsersQuery } from './users.query';
-// import { environment } from '$env';
-
+/** Configuration */
+interface StoreState extends EntityState<Models.User> {}
 const url = '//jsonplaceholder.typicode.com/users'; // environment.endpoints.apiUrl +
+const uniqueId = 'id';
 
+/** Store */
+@Injectable({ providedIn: 'root' })
+@StoreConfig({ name: 'users', idKey: uniqueId, resettable: true }) // , cache: { ttl: null }
+export class UsersStore extends EntityStore<StoreState, Models.User> {
+  constructor() {
+    super({ modifying: false, loading: false });
+  }
+}
+
+/** Query */
+// tslint:disable-next-line:max-classes-per-file
+@Injectable({ providedIn: 'root' })
+export class UsersQuery extends QueryEntity<StoreState, Models.User> {
+  public users$ = this.select();
+
+  constructor(protected store: UsersStore) {
+    super(store);
+  }
+}
+
+/** Service */
+// tslint:disable-next-line:max-classes-per-file
 @Injectable({ providedIn: 'root' })
 export class UsersService {
   constructor(
